Finished Blog

+++ admin/delete-category.php
@@ -5,15 +5,15 @@ if (isset($_GET['id'])) {
     $id = filter_var($_GET['id'], FILTER_SANITIZE_NUMBER_INT);
 
     //update category_id of posts deleted to default category
+    $update_query = "UPDATE posts SET category_id=5 WHERE  category_id=$id";
+    $update_result = mysqli_query($connection, $update_query);
 
-
-
-
-
-    //delete category
-    $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
-    $result = mysqli_query($connection, $query);
-    $_SESSION['delete-category-success'] = "Category deleted successfully";
+    if (!mysqli_errno($connection)) {
+        //delete category
+        $query = "DELETE FROM categories WHERE id=$id LIMIT 1";
+        $result = mysqli_query($connection, $query);
+        $_SESSION['delete-category-success'] = "Category deleted successfully";
+    }
 }
 header('location:' . ROOT_URL . 'admin/manage-categories.php');
 die();
